requirements.txt
asgiref~=3.5.0
Babel~=2.9.1
certifi~=2021.10.8
charset-normalizer~=2.0.12
Django~=4.0.2
django-modeladmin-reorder~=0.3.1
django-preferences~=1.0.0
emoji~=1.7.0
idna~=3.2
pytz~=2022.1
requests~=2.27.1
sqlparse~=0.4.1
urllib3~=1.26.9
pyTelegramBotAPI~=4.4.0
python-dotenv~=0.20.0
envparse~=0.2.0